fix(listing-skills): refactorisation part I
