+++ src/modules/skills/components/SkillCard.tsx
@@ -3,15 +3,15 @@ import { Button } from "@/shared/components/ui/button";
 import { DifficultyBadge } from "./DifficultyBadge";
 import { StatusBadge } from "./StatusBadge";
 import { ProgressBar } from "./ProgressBar";
-import { cn } from "../../../shared/utils/helpers";
+import { cn, formatDuration, formatDate } from "../../../shared/utils/helpers";
 import type { Skill } from "../skills.type";
 
-interface SkillCardProps {
+type SkillCardProps = {
   skill: Skill;
   onClick?: () => void;
   className?: string;
   style?: React.CSSProperties;
-}
+};
 
 export function SkillCard({
   skill,
@@ -19,23 +19,6 @@ export function SkillCard({
   className,
   style,
 }: SkillCardProps) {
-  const formatDuration = (minutes: number) => {
-    if (minutes < 60) return `${minutes} min`;
-    const hours = Math.floor(minutes / 60);
-    const remainingMinutes = minutes % 60;
-    return remainingMinutes > 0
-      ? `${hours}h ${remainingMinutes}min`
-      : `${hours}h`;
-  };
-
-  const formatDate = (dateString: string) => {
-    return new Date(dateString).toLocaleDateString("fr-FR", {
-      day: "numeric",
-      month: "short",
-      year: "numeric",
-    });
-  };
-
   return (
     <Card
       className={cn(
