+++ src/modules/skills/skills.const.ts
@@ -1,0 +1,17 @@
+export const difficultyConfig = {
+  Facile: {
+    icon: "🟢",
+    className:
+      "bg-green-100 text-green-800 border-green-200 dark:bg-green-900/20 dark:text-green-400 dark:border-green-800",
+  },
+  Moyen: {
+    icon: "🟡",
+    className:
+      "bg-yellow-100 text-yellow-800 border-yellow-200 dark:bg-yellow-900/20 dark:text-yellow-400 dark:border-yellow-800",
+  },
+  Difficile: {
+    icon: "🔴",
+    className:
+      "bg-red-100 text-red-800 border-red-200 dark:bg-red-900/20 dark:text-red-400 dark:border-red-800",
+  },
+};
